[refacto] renaming custom source into source scan
