+++ rust/datafusion/src/datasource/datasource.rs
@@ -17,11 +17,12 @@
 
 //! Data source traits
 
-use std::sync::Arc;
+use std::{collections::HashSet, fmt, sync::Arc};
 
 use crate::arrow::datatypes::SchemaRef;
 use crate::error::Result;
-use crate::physical_plan::ExecutionPlan;
+use crate::physical_plan::{ExecutionPlan, Partitioning, SendableRecordBatchStream};
+use async_trait::async_trait;
 
 /// Source table
 pub trait TableProvider {
@@ -34,4 +35,27 @@ pub trait TableProvider {
         projection: &Option<Vec<usize>>,
         batch_size: usize,
     ) -> Result<Arc<dyn ExecutionPlan>>;
+}
+
+#[async_trait]
+/// A scanner implementation that can be used by datafusion
+pub trait SourceScanner: Send + Sync + fmt::Debug {
+    /// reference to the schema of the data as it will be read by this scanner
+    fn projected_schema(&self) -> &SchemaRef;
+
+    /// string display of this scanner
+    fn format(&self) -> &str;
+
+    /// apply projection on this scanner
+    fn project(
+        &self,
+        required_columns: &HashSet<String>,
+        has_projection: bool,
+    ) -> Result<Arc<dyn SourceScanner>>;
+
+    /// get scanner partitioning
+    fn output_partitioning(&self) -> Partitioning;
+
+    /// get iterator for a given partition
+    async fn execute(&self, partition: usize) -> Result<SendableRecordBatchStream>;
 }
