+++ rust/datafusion/src/datasource/mod.rs
@@ -18,7 +18,6 @@
 //! DataFusion data sources
 
 pub mod csv;
-pub mod custom;
 pub mod datasource;
 pub mod memory;
 pub mod parquet;
