+++ rust/datafusion/src/physical_plan/mod.rs
@@ -202,7 +202,6 @@ pub mod aggregates;
 pub mod array_expressions;
 pub mod common;
 pub mod csv;
-pub mod custom_source;
 pub mod datetime_expressions;
 pub mod distinct_expressions;
 pub mod empty;
@@ -220,6 +219,7 @@ pub mod parquet;
 pub mod planner;
 pub mod projection;
 pub mod sort;
+pub mod source_scan;
 pub mod string_expressions;
 pub mod type_coercion;
 pub mod udaf;
